Add files via upload
This program is designed to run on Arduino, which is onboard the robot.
Connect to WiFi; set up a web server; Note: the home page contains buttons to drive robot either forward, backward, rotate on the spot, and move forward whist turning either left or right. Clicking buttons calls a web API. Web server (running from WiFi enabled robot) detects API calls and calls the appropriate functions to the drive the Sabertooth, and hence the motors.